Intro edits and Date and Time correction instructions
The date and time on the pi is often correct, and the reader needs to
know how to correct it when that happens.  Otherwise, the Pi won't
create valid SAS tokens when attempting to publish data to Azure.

The .csproj and .cs file changes are GitHub forcing line ending changes
on the files.  I don't seem to be able to avoid them.